separate the functionality
